+++ deploy/terraform/main.tf
@@ -1,0 +1,19 @@
+data "azurerm_client_config" "current" {}
+
+resource "random_string" "suffix" {
+  length  = 4
+  upper   = false
+  lower   = false
+  numeric = true
+  special = false
+}
+
+resource "azurerm_resource_group" "main" {
+  name     = var.resource_group_name
+  location = var.location
+  tags     = var.tags
+}
+
+locals {
+  acr_name = var.acr_name != "" ? var.acr_name : "acrdaedalus${random_string.suffix.result}"
+}